requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
async-timeout==5.0.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.2
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Compress==1.14
fonttools==4.57.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.4
numpy==2.0.2
oandapyV20==0.7.2
openai==1.74.0
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pydantic==2.11.3
pydantic_core==2.33.1
pyparsing==3.2.3
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.30
tomli==2.2.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
WTForms==3.2.1
zipp==3.21.0
gunicorn
plotly==5.18.0
nltk==3.8.1
rouge==1.0.1

# Testing
pytest==8.3.5
pytest-cov==2.12.1
playwright==1.38.0
pytest-playwright==0.4.0

# Monitoring and metrics
psutil==5.8.0
redis==5.2.1  # Updated to use consistent version
prometheus-client==0.11.0

# Security
PyJWT==2.3.0
cryptography==3.4.8

# Data visualization
plotly==5.18.0

# Performance and caching
cachetools==4.2.4
lru-cache==0.2.3

# Production server
gunicorn==20.1.0
gevent==24.2.1  # Updated to latest stable version
